+++ test/all_test.dart
@@ -4,10 +4,12 @@ import 'package:test/test.dart';
 
 import 'unit/iterable_test.dart' as unit_iterable;
 import 'unit/list_test.dart' as unit_list;
+import 'unit/string_test.dart' as unit_string;
 
 void main() {
   group('unit', () {
     unit_iterable.main();
     unit_list.main();
+    unit_string.main();
   });
 }