+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Added
   
 - Collection's classes and extensions
+- String extensions